creates route for climb list plus climbs component and renders data on the page
